transform get_attr_val() to return a double instead of an int, use -1.0 for "bus" attribute, otherwise use positive double values to indicate actual line widths. This allows specifying custom widths in the future, but still maintaining compatibility with bus=1. "bus" struct members are thus doubles now

+++ src/xschem.h
@@ -459,7 +459,7 @@ typedef struct
   short sel;
   char  *node;
   char *prop_ptr;
-  int bus; /*  20171201 cache here wire "bus" property, to avoid too many get_tok_value() calls */
+  double bus; /*  20171201 cache here wire "bus" property, to avoid too many get_tok_value() calls */
 } xWire;
 
 typedef struct
@@ -471,7 +471,7 @@ typedef struct
   unsigned short sel;
   char *prop_ptr;
   short dash;
-  int bus;
+  double bus;
 } xLine;
 
 #if HAS_CAIRO==1
@@ -519,7 +519,7 @@ typedef struct
   char *prop_ptr;
   short fill;
   short dash;
-  int bus;
+  double bus;
 } xPoly;
 
 typedef struct
@@ -1671,7 +1671,7 @@ extern char *my_strtok_r(char *str, const char *delim, const char *quote, int ke
 extern char **parse_cmd_string(const char *cmd, int *argc);
 extern int my_strncpy(char *d, const char *s, size_t n);
 extern int my_strcasecmp(const char *s1, const char *s2);
-extern int get_attr_val(const char *str);
+extern double get_attr_val(const char *str);
 extern int strboolcmp(const char *str, const char *boolean);
 extern char *my_strcasestr(const char *haystack, const char *needle);
 extern double mylog10(double x);
